rename tcc_prep file
it seems that github doesn't notice lower/uppercase changes?
